requirements.txt
flask == 2.0.3
flask-login == 0.5.0
flask-sqlalchemy == 2.5.1
requests == 2.27.1